% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ant_search.R
\docType{class}
\name{antsLog-class}
\alias{antsLog-class}
\title{Class: data structures related to the kernel of a funGp model}
\description{
Fill this!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{sols}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{args}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{fitness}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
