% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_outilsCode.R
\name{axtags}
\alias{axtags}
\title{Fitting of functional-input Gaussian process models}
\usage{
axtags(x)
}
\arguments{
\item{x}{Fill!!!!!!!!!!}
}
\description{
Creates a Gaussian process model based on the nature of the inputs which could be scalar,
functional or hybrid.
}
\keyword{internal}
