% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpMaster.R
\name{checkVal_funGp}
\alias{checkVal_funGp}
\title{Fill!!!!!!!!!!!}
\usage{
checkVal_funGp(env)
}
\arguments{
\item{env}{Fill!!!!!!!!!!!}
}
\description{
Fill!!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
