% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{checkVal_simulate}
\alias{checkVal_simulate}
\title{Fill!!!!!!!!!!!}
\usage{
checkVal_simulate(env)
}
\arguments{
\item{env}{Fill!!!!!!!!!!!}
}
\description{
Fill!!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
