% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGpFactory_Class.R
\docType{class}
\name{funGpFactory-class}
\alias{funGpFactory-class}
\title{Class: Fill!!!!}
\description{
Fill this!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{ppp}}{Object of class \code{"character"}. Fill!!!!!!!!!!}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
