% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ant_admin.R
\name{format4pred}
\alias{format4pred}
\title{Fitting of functional-input Gaussian process models}
\usage{
format4pred(sIn.pr = NULL, fIn.pr = NULL, args)
}
\arguments{
\item{sIn.pr}{Fill!!!!!!!!!!}

\item{fIn.pr}{Fill!!!!!!!!!!}

\item{args}{Fill!!!!!!!!!!}
}
\description{
Creates a Gaussian process model based on the nature of the inputs which could be scalar,
functional or hybrid. For functional inputs, the user might specify a projection method, projection
dimension and distance type seeking for optimal processing time or metamodel predictability.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
