% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_outilsCode.R
\name{formatShort}
\alias{formatShort}
\title{Put the "summary" table for a \code{Xfgpm} object in a
    shorter format using suitable abbreviations.}
\usage{
formatShort(structData)
}
\arguments{
\item{structData}{A data frame with character columns describing
the structural parameters of a \code{Xfgpm} object: state of
the variables (active or not), kernel used, ...}
}
\value{
A data frame with character columns corresponding to
    abbreviated versions of the content of \code{data}.
}
\description{
Abbreviate the column names and the column content of a data frame
describing the structure of \code{Xfgpm} object.
}
\note{
This function is not exported. It could work with a
    character matrix rather than a data frame. Mind that the input
    names of a \code{Xfgpm} object are implicitely assumed to be
    "X1", "X2", ... for the scalar inputs and "F1", "F2",
    ... while the matrix \code{sIn} and the list \code{fIn} given
    at the creation time may have different names!
}
\author{
Yves
}
