% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\docType{class}
\name{funGp-class}
\alias{funGp-class}
\title{Class: functional input Gaussian process model}
\description{
To create a funGp object, use \link[funGp]{funGp} . See also this function for mode details.
}
\section{Slots}{

\describe{
\item{\code{call}}{Object of class \code{"language"}. User call reminder.}

\item{\code{ds}}{Object of class \code{"numeric"}. Number of scalar inputs.}

\item{\code{df}}{Object of class \code{"numeric"}. Number of functional inputs.}

\item{\code{fDims}}{Object of class \code{"numeric"}. Dimension of each functional input.}

\item{\code{sIn}}{Object of class \code{"matrix"}. Scalar inputs.}

\item{\code{fIn}}{Object of class \code{"list"}. Functional inputs. Each element of the list contains a functional input in the form of a matrix.}

\item{\code{sOut}}{Object of class \code{"matrix"}. Scalar output.}

\item{\code{n.tr}}{Object of class \code{"integer"}. Number of training points.}

\item{\code{proj}}{Object of class \code{"funGpProj"}. Data structures related to the projection.}

\item{\code{kern}}{Object of class \code{"funGpKern"}. Data structures related to the kernel.}

\item{\code{preMats}}{Object of class \code{"list"}. L and LInvY matrices pre-computed for prediction. L is a lower diagonal matrix such that
\eqn{L'L} equals the training cross covariance matrix \eqn{K.tt}. On the other hand, \eqn{LInvY = L^(-1) * sOut}.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
