% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpMaster.R
\name{funGp}
\alias{funGp}
\title{Fitting of functional-input Gaussian process models}
\usage{
funGp(
  sIn = NULL,
  fIn = NULL,
  sOut,
  doProj = T,
  fpDims = NULL,
  kerType = "matern5_2",
  disType = "functional",
  n.starts = 1,
  n.presample = 1
)
}
\arguments{
\item{sIn}{a matrix of scalar input values to train the model. Each column must match an input variable and each row a training point.}

\item{fIn}{a list of functional inputs values to fit the model. Each element of the list must contain a matrix.}

\item{sOut}{a vector (or 1-column matrix) containing the values of the scalar output at the training points.}

\item{doProj}{a boolean indicating whether a projection of the inputs should be done for dimension reduction.}

\item{fpDims}{an optional array with the projection dimension for each functional input.}

\item{kerType}{an optional character specifying the covariance structure to be used. To be chosen between "gauss", "matern5_2" and
"matern3_2". Default is "matern5_2".}

\item{disType}{an optional character specifying the distance function to use for the functional inputs within the covariance
function. To be chosen between "scalar" and "functional". Default is "functional".}

\item{n.starts}{Fill!!!!!!!!!!}

\item{n.presample}{Fill!!!!!!!!!!}
}
\description{
Creates a Gaussian process model based on the nature of the inputs which could be scalar, functional or hybrid.
             For functional inputs, the user might specify a projection method, projection dimension and distance type seeking
             for optimal processing time or metamodel predictability.
}
\author{
José Betancourt
}
