% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_X-funGp_Class.R
\name{funGp_factory}
\alias{funGp_factory}
\title{Fill!!!}
\usage{
funGp_factory(
  sIn = NULL,
  fIn = NULL,
  sOut = NULL,
  ind.vl = NULL,
  method = "ACO",
  constraints = list(),
  setup = list()
)
}
\arguments{
\item{sIn}{fill}

\item{fIn}{fill}

\item{sOut}{fill}

\item{constraints}{fill}

\item{setup}{fill}
}
\description{
Fill. This.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
