% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_correlFunctions.R
\name{gaussian_cor}
\alias{gaussian_cor}
\title{Gaussian correlation function}
\usage{
gaussian_cor(Ms, thetas)
}
\arguments{
\item{Ms}{a list with as many elements as inputs. Each element of the list must be a n times n matrix of differences
between the observation coordinates, with n the number of input coordinates.}

\item{thetas}{an array with the length-scale parameters corresponding to the inputs for which the distance matrices were provided.}
}
\value{
A correlation matrix of dimension n times n computed from the Gaussian correlation function.
}
\description{
Applies the Gaussian correlation function to a list of distance matrices using corresponding length-scale parameters.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
