% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{getBasis}
\alias{getBasis}
\alias{getBasis,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{getBasis}{funGp}(object)
}
\arguments{
\item{object}{An object to predict from.}
}
\description{
This is my description
}
