% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Xfgpm_Class.R
\name{getFitness}
\alias{getFitness}
\title{Fitting of functional-input Gaussian process models}
\usage{
getFitness(model, sIn.vl = NULL, fIn.vl = NULL, sOut.vl = NULL, active = NULL)
}
\arguments{
\item{model}{Fill!!!!!!!!!!}

\item{sIn.vl}{Fill!!!!!!!!!!}

\item{fIn.vl}{Fill!!!!!!!!!!}

\item{sOut.vl}{Fill!!!!!!!!!!}

\item{active}{Fill!!!!!!!!!!}
}
\description{
Creates a Gaussian process model based on the nature of the inputs which could be scalar,
functional or hybrid. For functional inputs, the user might specify a projection method, projection
dimension and distance type seeking for optimal processing time or metamodel predictability.
}
\keyword{internal}
