% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpModel_Class.R
\name{getGram}
\alias{getGram}
\alias{getGram,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{getGram}{funGp}(object)
}
\arguments{
\item{object}{An object to predict from.}
}
\description{
This is my description
}
