% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGp_Class.R
\name{getProjfIn}
\alias{getProjfIn}
\alias{getProjfIn,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{getProjfIn}{funGp}(object)
}
\arguments{
\item{object}{An object to predict from.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
