% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraDoc.R, R/funGpModel_Class.R
\name{getProjfuns}
\alias{getProjfuns}
\alias{getProjfuns,funGp-method}
\title{List of projected functional inputs}
\usage{
\S4method{getProjfuns}{funGp}(object)
}
\arguments{
\item{object}{a funGp model. Might be either a funGp_S, funGp_F or funGp_SF object depending on the type
of model built.}
}
\description{
Provides a list with the functional inputs rebuilt from the basis functions and projection
             coefficients used for dimension reduction.
}
\author{
José Betancourt
}
