% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraDoc.R, R/funGpModel_Class.R
\name{getProjgram}
\alias{getProjgram}
\alias{getProjgram,funGp-method}
\title{List of gram matrices}
\usage{
\S4method{getProjgram}{funGp}(object)
}
\arguments{
\item{object}{a funGp model. Might be either a funGp_S, funGp_F or funGp_SF object depending on the type
of model built.}
}
\description{
Provides a list with the gram matrices built from the basis functions used to project each
             functional input
}
\author{
José Betancourt
}
