% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{getTrainCov}
\alias{getTrainCov}
\alias{getTrainCov,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{getTrainCov}{funGp}(object)
}
\arguments{
\item{object}{An object to predict from.}
}
\description{
This is my description
}
