% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_correlFunctions.R
\name{matern32_cor}
\alias{matern32_cor}
\title{Matern 3/2 correlation function}
\usage{
matern32_cor(Ms, thetas)
}
\arguments{
\item{Ms}{a list with as many elements as inputs. Each element of the list must be a n times n matrix of differences
between the observation coordinates, with n the number of input coordinates.}

\item{thetas}{an array with the length-scale parameters corresponding to the inputs for which the distance matrices were provided.}
}
\value{
A correlation matrix of dimension n times n computed from the Matern 3/2 correlation function.
}
\description{
Applies the Matern 5/2 correlation function to a list of distance matrices using corresponding length-scale parameters.
}
\author{
José Betancourt, François Bachoc, Thierry Klein and Jérémy Rohmer
}
\keyword{internal}
