% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_outilsCode.R
\docType{class}
\name{modelCall-class}
\alias{modelCall-class}
\title{S4 class for calls to the fgpm function in funGp}
\description{
User reminder of the \link[funGp]{fgpm} function call.
}
\section{Slots}{

\describe{
\item{\code{string}}{Object of class \code{"character"}. User call reminder in string format.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
