% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_S.R
\name{negLogLik_funGp_S}
\alias{negLogLik_funGp_S}
\title{Setting up of starting points for optimization of hyperparameters}
\usage{
negLogLik_funGp_S(thetas_s, sMs, sOut, kerType, varfun, nugget)
}
\arguments{
\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}

\item{theta_s}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
