% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_SF.R
\name{negLogLik_funGp_SF}
\alias{negLogLik_funGp_SF}
\title{Setting up of starting points for optimization of hyperparameters}
\usage{
negLogLik_funGp_SF(
  thetas,
  sMs,
  fMs,
  sOut,
  kerType,
  varfun,
  ls_s.known,
  ls_f.known,
  nugget
)
}
\arguments{
\item{thetas}{Fill this!!!!!!!!!!}

\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
