% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_F.R
\name{optimHypers_F}
\alias{optimHypers_F}
\title{Optimizing the hyperparameters}
\usage{
optimHypers_F(
  spoints,
  n.starts,
  bnds,
  fMs,
  sOut,
  kerType,
  varfun,
  nugget,
  par.clust
)
}
\arguments{
\item{spoints}{Fill this!!!!!!!!!!}

\item{n.starts}{Fill this!!!!!!!!!!}

\item{bnds}{Fill this!!!!!!!!!!}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
