% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_S.R
\name{optimHypers_S}
\alias{optimHypers_S}
\title{Optimizing the hyperparameters}
\usage{
optimHypers_S(
  spoints,
  n.starts,
  bnds,
  sMs,
  sOut,
  kerType,
  varfun,
  nugget,
  par.clust
)
}
\arguments{
\item{spoints}{Fill this!!!!!!!!!!}

\item{bnds}{Fill this!!!!!!!!!!}

\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
