% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plot}
\alias{plot}
\alias{plot,Xfgpm-method}
\title{Plot a \code{Xfgpm} object}
\usage{
plot(x, y, ...)

\S4method{plot}{Xfgpm}(x, y = NULL, which = c("evol", "diag"), calib = TRUE, fitp = TRUE, ...)
}
\arguments{
\item{x}{the \code{Xfgpm} object to plot.}

\item{y}{not used.}

\item{...}{Other graphical parameters to be passed to
\code{plotEvol} or \code{plotX} such as \code{main} of
\code{xlab}. When \code{which} is \code{"diag"} and both
\code{calib} and \code{fitp} are \code{TRUE}, the graphical
should be enclosed into a list and passed with the formal name
\code{calib.gpars} or \code{fitp.gpars}.}

\item{which}{character giving the type of plot wanted. When the
value is \code{"evol"} the method \code{\link{plotEvol}} is
used; when the value is \code{"diag"} the method
\code{\link{plotX}} is used.}

\item{calib}{logical. If \code{TRUE}}

\item{fitp}{logical. If \code{TRUE}.}
}
\value{
Nothing.
}
\description{
Plot an object with class \code{"Xfgpm"}.
}
\note{
This method is provided only. See \code{\link{plotEvol}} and
    \code{\link{plotX}} methods.
}
