% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctionsMore.R
\name{plot,Xfgpm-method}
\alias{plot,Xfgpm-method}
\title{Plot method for the class \code{"Xfgpm"}}
\usage{
\S4method{plot}{Xfgpm}(
  x,
  y = NULL,
  which = c("evol", "diag"),
  calib = TRUE,
  fitp = TRUE,
  horiz = FALSE,
  ...
)
}
\arguments{
\item{x}{the \code{Xfgpm} object to plot.}

\item{y}{not used.}

\item{which}{character giving the type of plot wanted. When the
value is \code{"evol"} the method \code{\link{plotEvol}} is
used; when the value is \code{"diag"} the method
\code{\link{plotX}} is used. See \bold{Examples}.}

\item{calib}{logical. If \code{TRUE} the calibration plot of the
selected model will be included in the display in its
"diagnostic" part if \code{which} is set to \code{"diag"}.}

\item{fitp}{logical. If \code{TRUE} a scatter plot of the quality
of all explored models will be included in the display in its
"diagnostic" part if \code{which} is set to \code{"diag"}.}

\item{horiz}{logical. Used only when \code{which} is \code{"diag"}
and when both \code{calib} and \code{fitp} are \code{TRUE}. If
\code{horiz} is \code{TRUE} the two subplots are displayed in
horizontally (on a same row) rather than vertically which is
the default.}

\item{...}{Other graphical parameters to be passed to
\code{plotEvol} or \code{plotX} such as \code{main} of
\code{xlab}. When \code{which} is \code{"diag"} and both
\code{calib} and \code{fitp} are \code{TRUE}, the graphical
should be enclosed into a list and passed with the formal name
\code{calib.gpars} or \code{fitp.gpars}.}
}
\value{
Nothing.
}
\description{
Plot an object with class \code{"Xfgpm"} representing
    a collection of functional GP models corresponding to
    different structural parameters.

Two types of graphics can be shown depending on the choice of
\code{which}. The choice \code{which = "evolution"} is used to
assess the quality of the fitted \code{fgpm} models on the basis
of Leave-One-Out cross-validation.  The choice \code{which =
"diag"} is used to display diagnostics. Two types of diagnostic
plots are shown as sub-plots by default, but each can be discarded
if wanted.
}
\note{
This method is provided only XXXY. See \code{\link{plotEvol}} and
    \code{\link{plotX}} methods.
}
\examples{
# generating input and output data
set.seed(100)
n.tr <- 2^5
x1 <- x2 <- x3 <- x4 <- x5 <- seq(0, 1, length = n.tr^(1/5))
sIn <- expand.grid(x1 = x1, x2 = x2, x3 = x3, x4 = x4, x5 = x5)
fIn <- list(f1 = matrix(runif(n.tr * 10), ncol = 10),
            f2 = matrix(runif(n.tr * 22), ncol = 22))
sOut <- fgp_BB7(sIn, fIn, n.tr)
\dontrun{
# optimizing the model structure with 'fgpm_factory' (~5 seconds)
xm <- fgpm_factory(sIn = sIn, fIn = fIn, sOut = sOut)
# assessing the quality of the model - absolute and w.r.t. the other
# explored models
plot(xm, which = "evol")
# Diagnostics (two subplots)
plot(xm, which = "diag")
plot(xm, which = "diag", horiz = TRUE)
# Diagnostics (one plot)
plot(xm, which = "diag", fitp = FALSE)
plot(xm, which = "diag", calib = FALSE)
}
}
