% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{plotCalib}
\alias{plotCalib}
\alias{plotCalib,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotCalib}{funGp}(object, preds, sOut.pr, ...)
}
\arguments{
\item{object}{An object to predict from.}

\item{preds}{something}

\item{sOut.pr}{also}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
