% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotLAA,fgpm-method}
\alias{plotLAA,fgpm-method}
\title{Leave-one-out calibration plot for a funGp model}
\usage{
\S4method{plotLAA}{fgpm}(model, ...)
}
\description{
This method provides a diagnostic plot for the validation of a funGp Gaussian process model.
  It displays a calibration plot based on the leave-one-out predictions of the output at the points
  used to train the model.
}
\keyword{internal}
