% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotLAA}
\alias{plotLAA}
\title{Leave-one-out calibration plot for a funGp model}
\usage{
plotLAA(model, ...)
}
\arguments{
\item{model}{an object of class \linkS4class{fgpm} corresponding to the funGp model to validate.}

\item{...}{additional arguments affecting the plot. The following typical graphics parameters are
valid entries: \emph{xlim}, \emph{ylim}, \emph{xlab}, \emph{ylab}, \emph{main}.}
}
\description{
This method provides a diagnostic plot for the validation of a funGp Gaussian process model.
  It displays a calibration plot based on the leave-one-out predictions of the output at the points
  used to train the model.
}
