% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotLOO}
\alias{plotLOO}
\alias{plotLOO,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotLOO}{funGp}(model, ...)
}
\arguments{
\item{model}{An object to predict from.}

\item{...}{fill}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
