% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotPreds}
\alias{plotPreds}
\alias{plotPreds,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotPreds}{funGp}(model, preds, sOut.pr = NULL, calib = T, legends = T, stat = T, ...)
}
\arguments{
\item{model}{fill}

\item{preds}{fill}

\item{sOut.pr}{also}

\item{calib}{also}

\item{legends}{also}

\item{stat}{also}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
