% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{plotSims}
\alias{plotSims}
\alias{plotSims,funGp-method}
\title{Fill Method for the apk Class}
\usage{
\S4method{plotSims}{funGp}(object, sims = sims, detail = "full", ...)
}
\arguments{
\item{object}{An object to predict from.}

\item{sims}{something}

\item{detail}{fill!!!!}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
