% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotX}
\alias{plotX}
\alias{plotX,XfunGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotX}{XfunGp}(x.model, ...)
}
\arguments{
\item{x.model}{An object to predict from.}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
