% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpModel_Class.R
\name{predict}
\alias{predict}
\alias{predict.funGpModel}
\title{predict for \code{funGpModel} object}
\usage{
\method{predict}{funGpModel}(object, sIn.pr = NULL, fIn.pr = NULL, detail = "light", ...)
}
\arguments{
\item{object}{A \code{funGpModel} object}

\item{sIn.pr}{Fill!!!!!!!!!!}

\item{fIn.pr}{Fill!!!!!!!!!!}

\item{detail}{Fill!!!!!!!!!!}

\item{...}{Additional arguments}
}
\description{
predict for \code{funGpModel} object
}
