% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGpProj_Class.R
\docType{class}
\name{funGpProj-class}
\alias{funGpProj-class}
\title{Class: data structures related to projection of functional inputs}
\description{
Fill this!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{doProj}}{Object of class \code{"logical"}. Should projection of functional inputs be done?}

\item{\code{fpDims}}{Object of class \code{"numeric"}. Projection dimension of each functional input.}

\item{\code{basis}}{Object of class \code{"list"}. Each element (fDims_i x fpDims_i) contains the basis
functions used for the projection of one functional input.}

\item{\code{coefs}}{Object of class \code{"list"}. Each element (n x fpDims_i) contains the coefficients
used for the projection of one functional input.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
