% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_F.R
\name{setBounds_F}
\alias{setBounds_F}
\title{Setting up of hyperparameters' boundaries for optimization}
\usage{
setBounds_F(fMs)
}
\arguments{
\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}
}
\value{
A matrix with two rows and as many columns as hyperparameters the metamodel requires. The first row indicates the lower limits
and the second row the upper limits for the optimization.
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
