% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_S.R
\name{setBounds_S}
\alias{setBounds_S}
\title{Setting up of hyperparameters' boundaries for optimization}
\usage{
setBounds_S(sMs)
}
\arguments{
\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}
}
\value{
A matrix with two rows and as many columns as hyperparameters the metamodel requires. The first row indicates the lower limits
        and the second row the upper limits for the optimization.
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
