% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_SF.R
\name{setBounds_SF}
\alias{setBounds_SF}
\title{Setting up of hyperparameters' boundaries for optimization}
\usage{
setBounds_SF(sMs, fMs)
}
\arguments{
\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}
}
\value{
A matrix with two rows and as many columns as hyperparameters the metamodel requires. The first row indicates the lower limits
and the second row the upper limits for the optimization.
}
\description{
Fill this!!!!!!!!!!
}
\author{
Jose Betancourt
}
\keyword{internal}
