% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_distanceFunctions.R
\name{setFunDistance}
\alias{setFunDistance}
\title{Computation of distance matrix for functional inputs}
\usage{
setFunDistance(fpIn1, fpIn2, J)
}
\arguments{
\item{fpIn1}{a first list with as many elements as functional inputs. The i-th element must be a matrix with the projection coefficients
for the i-th functional input.}

\item{fpIn2}{a second list with as many elements as functional inputs. The i-th element must be a matrix with the projection coefficients
for the i-th functional input.}

\item{J}{a list with as many elements as functional inputs. The i-th element must be the Gram matrix of the basis functions used for
the projection of the i-th functional input.}
}
\value{
A list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.
}
\description{
Precomputes the distance matrices that will be used later inside the \emph{functional} term of the kernel function
             during the optimization of the hyperparmeters of the metamodel.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
