% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_F.R
\name{setHypers_F}
\alias{setHypers_F}
\title{Optimization of hyperparameters for funGp models}
\usage{
setHypers_F(
  fpIn,
  J,
  fMs,
  sOut,
  kerType,
  var.known,
  ls_f.known,
  n.starts,
  n.presample
)
}
\arguments{
\item{fpIn}{a list with as many elements as functional inputs. The i-th element must be a matrix with the projection coefficients
for the i-th functional input.}

\item{J}{a list with as many elements as functional inputs. The i-th element must be the Gram matrix of the basis functions used for
the projection of the i-th functional input.}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{a vector (or 1-column matrix) containing the values of the scalar output at the training points.}

\item{kerType}{a character string indicating the covariance structure to be used, to be choosen between "gauss", "matern5_2" or "matern3_2".}

\item{var.known}{Fill!!!!!!!!!!}

\item{ls_f.known}{Fill!!!!!!!!!!}

\item{n.starts}{Fill!!!!!!!!!!}

\item{n.presample}{Fill!!!!!!!!!!}
}
\value{
The hyperparameters.
}
\description{
Sets good values for the hyperparameters of the Gaussian process model based on optimization.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
