% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_SF.R
\name{setHypers_SF}
\alias{setHypers_SF}
\title{Optimization of hyperparameters for funGp models}
\usage{
setHypers_SF(
  sMs,
  fMs,
  sOut,
  kerType,
  var.known,
  ls_s.known,
  ls_f.known,
  n.starts,
  n.presample,
  nugget
)
}
\arguments{
\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{a vector (or 1-column matrix) containing the values of the scalar output at the training points.}

\item{kerType}{a character string indicating the covariance structure to be used, to be choosen between "gauss", "matern5_2" or "matern3_2".}

\item{var.known}{Fill!!!!!!!!!!}

\item{ls_s.known}{Fill!!!!!!!!!!}

\item{ls_f.known}{Fill!!!!!!!!!!}

\item{n.starts}{Fill!!!!!!!!!!}

\item{n.presample}{Fill!!!!!!!!!!}

\item{nugget}{Fill!!!!!!!!!!}
}
\value{
The hyperparameters.
}
\description{
Sets good values for the hyperparameters of the Gaussian process model based on optimization.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
