% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlFunctions.R
\name{setR}
\alias{setR}
\title{Fill this!!!}
\usage{
setR(thetas, Ms, kerType)
}
\arguments{
\item{thetas}{Fill this!!!!!!!!!!}

\item{Ms}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{kerType}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
