% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_distanceFunctions.R
\name{setScalDistance}
\alias{setScalDistance}
\title{Computation of distance matrix for scalar inputs}
\usage{
setScalDistance(sIn1, sIn2)
}
\arguments{
\item{sIn1}{a first matrix of scalar input values.}

\item{sIn2}{a second matrix of scalar input values.}
}
\value{
A list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.
}
\description{
Precomputes the distance matrices that will be used later inside the \emph{scalar} term of the kernel function
             during the optimization of the hyperparmeters of the metamodel.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
