% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_S.R
\name{setSPoints_S}
\alias{setSPoints_S}
\title{Setting up of starting points for optimization of hyperparameters}
\usage{
setSPoints_S(bnds, sMs, sOut, kerType, varfun, n.starts, n.presample)
}
\arguments{
\item{bnds}{Fill this!!!!!!!!!!}

\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}

\item{niter}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
