% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpKern_Class.R, R/funGpProj_Class.R,
%   R/funGp_Class.R
\name{show}
\alias{show}
\alias{show,funGpKern-method}
\alias{show,funGpProj-method}
\alias{show,funGp-method}
\title{Fill!!!!!!!!!!!}
\usage{
\S4method{show}{funGpKern}(object)

\S4method{show}{funGpProj}(object)

\S4method{show}{funGp}(object)
}
\arguments{
\item{object}{An object to show.}
}
\description{
This is my description

This is my description

This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein

José Betancourt, François Bachoc and Thierry Klein

José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
