% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraDoc.R
\name{show.funGpModel}
\alias{show.funGpModel}
\title{Printing of a funGp model}
\arguments{
\item{object}{a funGp model. Might be either a funGp_S, funGp_F or funGp_SF object depending on the type
of model built.}
}
\description{
Prints the main features of a funGp model. The list of attributes printed depends on the
             type of model built, which can be scalar-input, functional-input or hybrid-input.
}
\author{
José Betancourt
}
