% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{simulate}
\alias{simulate}
\alias{simulate,funGp-method}
\title{Simulation method for the funGp Class}
\usage{
simulate(object, nsim = 1, seed = NULL, ...)

\S4method{simulate}{funGp}(
  object,
  nsim = 1,
  seed = NULL,
  sIn.sm = NULL,
  fIn.sm = NULL,
  nug.sim = 0,
  detail = "light",
  ...
)
}
\arguments{
\item{object}{An object to simulate from.}

\item{nsim}{fill!!!}

\item{seed}{fill!!!}

\item{...}{Further arguments for methods.}

\item{sIn.sm}{fill!!}

\item{fIn.sm}{fill!!}

\item{nug.sim}{fill!!!}

\item{detail}{fill!!!}
}
\description{
Simulation method for the funGp Class
}
\examples{
# generating input data for training
n.tr <- 25
sIn.tr <- expand.grid(x1 = seq(0,1,length = sqrt(n.tr)), x2 = seq(0,1,length = sqrt(n.tr)))
sIn.tr <- as.matrix(sIn.tr)
fIn.tr <- list(f1 = matrix(runif(n.tr*10), ncol = 10), matrix(runif(n.tr*22), ncol = 22))

# generating output data for training
sOut.tr <- as.matrix(sapply(t(1:n.tr), function(i){
  x1 <- sIn.tr[i,1]
  x2 <- sIn.tr[i,2]
  f1 <- fIn.tr[[1]][i,]
  f2 <- fIn.tr[[2]][i,]
  as.numeric(x1 * sin(x2) + x1 * mean(f1) - x2^2 * diff(range(f2)))
}))

# creating a funGp model
m1 <- funGp(sIn = sIn.tr, fIn = fIn.tr, sOut = sOut.tr)

# generating input data for simulation
n.sm <- 100
sIn.sm <- expand.grid(x1 = seq(0,1,length = sqrt(n.sm)), x2 = seq(0,1,length = sqrt(n.sm)))
sIn.sm <- as.matrix(sIn.sm)
fIn.sm <- list(f1 = matrix(runif(n.sm*10), ncol = 10), matrix(runif(n.sm*22), ncol = 22))

# making simulations
m1.sims <- simulate(m1, nsim = 10, sIn.sm = sIn.sm, fIn.sm = fIn.sm, detail = "full")
matplot(t(m1.sims$obs), type = 'l', col = "grey")
lines(m1.sims$mean, col = "red")
lines(m1.sims$lower95, col = "blue")
lines(m1.sims$upper95, col = "blue")

}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
