% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_fgpm_Class.R
\name{summary,fgpm-method}
\alias{summary,fgpm-method}
\title{Summary method for \code{fgpm} objects}
\usage{
\S4method{summary}{fgpm}(object, ...)
}
\arguments{
\item{object}{An \code{fgpm} object.}

\item{...}{Not used yet.}
}
\description{
Display the structure of a \code{fgpm}
    object and the value of the parameters (variance and length-scales).
}
\note{
This method is actually identical to the \code{show} method
    for this class which is called when the name of the object is
    entered in an interactive session.
}
\examples{
m <- xm@model
class(m)
summary(m)
m
}
