% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_summary_Doc.R, R/1_fgpm_Class.R,
%   R/1_Xfgpm_Class.R
\name{summary}
\alias{summary}
\alias{summary,fgpm-method}
\alias{summary,Xfgpm-method}
\title{Summary methods for the funGp package}
\usage{
summary(object, ...)

\S4method{summary}{fgpm}(object)

\S4method{summary}{Xfgpm}(object, n = 24)
}
\arguments{
\item{object}{an \code{Xfgpm} object.}

\item{...}{Not used yet.}

\item{n}{maximal number of lines (\code{fgpm} objects) to show.}
}
\value{
an object inheriting from \code{data.frame}.
}
\description{
This set of methods produce summaries for the main
    \linkS4class{fgpm} and \linkS4class{Xfgpm} objects representing
    funGp models and structures related to structural optimization
    of the model, respectively.

Display a summary of the structure for up to \code{n} \code{fgpm}
objects visited during the ACO optimization.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
