% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGpFactory_Class.R
\name{symbolicfunGp}
\alias{symbolicfunGp}
\title{Also Fill!!}
\usage{
symbolicfunGp(sIn = NULL, fIn = NULL, sOut, tr.ind = NULL)
}
\arguments{
\item{sIn}{a matrix of scalar input values to train the model. Each column must match an input variable}

\item{fIn}{a list of functional inputs values to fit the model. Each element of the list must contain a
matrix.}

\item{sOut}{a vector (or 1-column matrix) containing the values of the scalar output at the training
points.}

\item{tr.ind}{matrix of training indices.}
}
\description{
Fill!!.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
