% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test3}
\alias{test3}
\title{MFR_1
First analytical example in Muehlenstaedt, Fruth & Roustant (2016)}
\usage{
test3(sIn, fIn, n.tr)
}
\description{
MFR_1
First analytical example in Muehlenstaedt, Fruth & Roustant (2016)
}
