% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test4}
\alias{test4}
\title{MFR_2p
Second analytical example in preprint of Muehlenstaedt, Fruth & Roustant (2016)}
\usage{
test4(sIn, fIn, n.tr)
}
\description{
MFR_2p
Second analytical example in preprint of Muehlenstaedt, Fruth & Roustant (2016)
}
