% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test5}
\alias{test5}
\title{MFR_2f
Second analytical example in final version of Muehlenstaedt, Fruth & Roustant (2016)}
\usage{
test5(sIn, fIn, n.tr)
}
\description{
MFR_2f
Second analytical example in final version of Muehlenstaedt, Fruth & Roustant (2016)
}
