% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test6}
\alias{test6}
\title{NHMPP
Function inspired in the analytical example in Nanty, Helbert, Marrel, Pérot, Prieur (2016)}
\usage{
test6(sIn, fIn, n.tr)
}
\description{
NHMPP
Function inspired in the analytical example in Nanty, Helbert, Marrel, Pérot, Prieur (2016)
}
