% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGp_Class.R
\name{update}
\alias{update}
\alias{update,funGp-method}
\title{Updating method for the funGp Class}
\usage{
update(object, ...)

\S4method{update}{funGp}(
  object,
  sIn.nw = NULL,
  fIn.nw = NULL,
  sOut.nw = NULL,
  sIn.sb = NULL,
  fIn.sb = NULL,
  sOut.sb = NULL,
  ind.sb = NULL,
  ind.dl = NULL,
  var.sb = NULL,
  ls_s.sb = NULL,
  ls_f.sb = NULL,
  var.re = F,
  ls_s.re = F,
  ls_f.re = F,
  ...
)
}
\arguments{
\item{object}{An object to update.}

\item{...}{Further arguments for methods.}

\item{sIn.nw}{Fill!!!}

\item{fIn.nw}{Fill!!!}

\item{sOut.nw}{Fill!!!}

\item{sIn.sb}{Fill!!!}

\item{fIn.sb}{Fill!!!}

\item{sOut.sb}{Fill!!!}

\item{ind.sb}{Fill!!!}

\item{ind.dl}{Fill!!!}

\item{var.sb}{Fill!!!}

\item{ls_s.sb}{Fill!!!}

\item{ls_f.sb}{Fill!!!}

\item{var.re}{Fill!!!}

\item{ls_s.re}{Fill!!!}

\item{ls_f.re}{Fill!!!}
}
\description{
Updating method for the funGp Class
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
