% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_XfunGp_Class.R
\docType{class}
\name{XfunGp-class}
\alias{XfunGp-class}
\title{Class: data structures related to the kernel of a funGp model}
\description{
Fill this!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{model}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{fitness}}{Hola.}

\item{\code{structure}}{Hola.}

\item{\code{factoryCall}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{n.solspace}}{Hola.}

\item{\code{n.explored}}{Hola.}

\item{\code{method}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{stat}}{Hola.}

\item{\code{details}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{log.success}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{log.crashes}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
