% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ant_admin.R
\docType{class}
\name{antsLog-class}
\alias{antsLog-class}
\title{S4 class for log of models explored by ant colony in funGp}
\description{
Register of model structures and their performance statistic, if available.
}
\section{Slots}{

\describe{
\item{\code{sols}}{Object of class \code{"data.frame"}. Compendium of model structures arranged by rows. Each
column is linked to one structural parameter of the model such as the state of one variable (inactive,
active) or the type of kernel function.}

\item{\code{args}}{Object of class \code{"list"}. Compendium of model structures represented by objects of class
\code{"\linkS4class{modelCall}"}}

\item{\code{fitness}}{Object of class \code{"numeric"}. Performance statistic of each model, if available.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
