% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpKern_Class.R
\docType{class}
\name{funGpKern-class}
\alias{funGpKern-class}
\title{Class: data structures related to the kernel of a funGp model}
\description{
Fill this!!!!!!!!!
}
\section{Slots}{

\describe{
\item{\code{kerType}}{Object of class \code{"character"}. Kernel type. To be chosen from {"gauss", "matern5_2", "matern3_2"}.}

\item{\code{disType}}{Object of class \code{"character"}. Distance type. To be chosen from {"scalar", "functional"}.}

\item{\code{varHyp}}{Object of class \code{"numeric"}. Estimated variance parameter.}

\item{\code{s_lsHyps}}{Object of class \code{"numeric"}. Estimated length-scale parameters for scalar inputs.}

\item{\code{f_lsHyps}}{Object of class \code{"numeric"}. Estimated length-scale parameters for functional inputs.}
}}

\author{
José Betancourt, François Bachoc and Thierry Klein
}
