% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_F.R
\name{negLogLik_funGp_F}
\alias{negLogLik_funGp_F}
\title{Setting up of starting points for optimization of hyperparameters}
\usage{
negLogLik_funGp_F(thetas_f, fMs, sOut, kerType)
}
\arguments{
\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}

\item{thetas}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt
}
\keyword{internal}
