% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{plotLOO}
\alias{plotLOO}
\alias{plotLOO,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotLOO}{funGp}(object, xlim = NULL, ylim = NULL)
}
\arguments{
\item{object}{An object to predict from.}

\item{xlim}{something}

\item{ylim}{something}

\item{...}{fill}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
