% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{plotPreds}
\alias{plotPreds}
\alias{plotPreds,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotPreds}{funGp}(object, preds, sOut.pr = NULL, ...)
}
\arguments{
\item{object}{An object to predict from.}

\item{preds}{something}

\item{sOut.pr}{also}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
