% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_funGp_Class.R
\name{plotPreds}
\alias{plotPreds}
\alias{plotPreds,funGp-method}
\title{Prediction Method for the apk Class}
\usage{
\S4method{plotPreds}{funGp}(
  object,
  preds,
  sOut.pr = NULL,
  xlim_c = NULL,
  ylim_c = NULL,
  justCal = F,
  justLin = F,
  ...
)
}
\arguments{
\item{object}{An object to predict from.}

\item{preds}{something}

\item{sOut.pr}{also}

\item{xlim_c}{also}

\item{ylim_c}{also}

\item{justCal}{also}

\item{justLin}{also}

\item{...}{Further arguments for methods.}
}
\description{
This is my description
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
