% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plottingFunctions.R
\name{plotX}
\alias{plotX}
\alias{plotX,Xfgpm-method}
\title{Diagnostic plots for funGp factory output}
\usage{
\S4method{plotX}{Xfgpm}(x.model, calib = TRUE, fitp = TRUE, ...)
}
\arguments{
\item{x.model}{an object of class \linkS4class{Xfgpm} containing the output of the model selection
algorithm in \link[funGp]{fgpm_factory}.}

\item{calib}{a boolean indicating wether the calibration plot of the selected model shoud be included
in the display. Default is TRUE.}

\item{fitp}{a boolean indicating wether scatter plot of the quality of all explored models should be
included in the display. Default is TRUE.}

\item{...}{additional arguments affecting the display. Since this method allows to generate two plots
from a single function call, the extra arguments for each plot should be included in a list. For the
calibration plot, the list should be called \emph{calib.gpars}. For the plot of the fitness of
explored models, the list should be called \emph{fitp.gpars}. The following typical graphics parameters
are valid entries of both lists: \emph{xlim}, \emph{ylim}, \emph{xlab}, \emph{ylab}, \emph{main}. The
boolean argument legends can also be included in any of the two lists in order to control the display
of legends in the corresponding plot.}
}
\description{
This method provides two plots for assessing the quality of the output delivered by the
  model selection algorithm in the \link[funGp]{fgpm_factory} function. The first one is a calibration
  plot similar to the one offered for \linkS4class{fgpm} objects by the \link[funGp]{plotLOO} function.
  This plot allows to validate the absolute quality of the selected model. The second one displays the
  performance statistic of all the models successfully evaluated by the model selection algorithm. This
  provides a notion of the relative quality of the selected model with respect to the other models that
  can be made using the same data.
}
\examples{
# generating input and output data
set.seed(100)
n.tr <- 2^5
sIn <- expand.grid(x1 = seq(0,1,length = n.tr^(1/5)), x2 = seq(0,1,length = n.tr^(1/5)),
                   x3 = seq(0,1,length = n.tr^(1/5)), x4 = seq(0,1,length = n.tr^(1/5)),
                   x5 = seq(0,1,length = n.tr^(1/5)))
fIn <- list(f1 = matrix(runif(n.tr*10), ncol = 10), f2 = matrix(runif(n.tr*22), ncol = 22))
sOut <- fgp_BB7(sIn, fIn, n.tr)
\donttest{
# optimizing the model structure with fgpm_factory (~5 seconds)
xm <- fgpm_factory(sIn = sIn, fIn = fIn, sOut = sOut)

# assessing the quality of the model - absolute and w.r.t. the other explored models
plotX(xm)

# customizing some graphical parameters
plotX(xm, calib.gpars = list(xlim = c(800,1000), ylim = c(600,1200)),
          fitp.gpars = list(main = "Relative quality", legends = F))
}

}
\references{
Betancourt, J., Bachoc, F., Klein, T., and Gamboa, F. (2020),
Technical Report: "Ant Colony Based Model Selection for Functional-Input Gaussian Process Regression. Ref. B3D-WP3.2".
\emph{RISCOPE project}.
\href{https://hal.archives-ouvertes.fr/hal-02532713}{[HAL]}

Betancourt, J., Bachoc, F., and Klein, T. (2020),
R Package Manual: "Gaussian Process Regression for Scalar and Functional Inputs with funGp - The in-depth tour".
\emph{RISCOPE project}.
\href{https://drive.google.com/file/d/1MtYi-Qq-BNZpbp1SWWG4Fb35HvVqRHQM/view?usp=sharing}{[HAL]}
}
\seealso{
\strong{*} \link[funGp]{fgpm_factory} for structural optimization of funGp models;

\strong{*} \link[funGp]{plotEvol} for a plot on the evolution of the model selection algorithm
  in fgpm_factory.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
