% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpModel_Class.R
\name{predict}
\alias{predict}
\alias{predict,funGpModel-method}
\title{Prediction method for the funGp Class}
\usage{
predict(object, ...)

\S4method{predict}{funGpModel}(object, ...)
}
\arguments{
\item{object}{An object to predict from.}

\item{...}{Further arguments for methods.}
}
\description{
Prediction method for the funGp Class
}
