% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fgpProj_Class.R
\docType{class}
\name{fgpProj-class}
\alias{fgpProj-class}
\title{S4 class for structures linked to projections in a \code{fgpm} model}
\description{
This is the formal representation for data structures linked to projection of inputs in a
  Gaussian process model within the \link[=funGp-package]{funGp package}.
}
\section{Slots}{

\describe{
\item{\code{pdims}}{Object of class \code{"numeric"}. Projection dimension of each input.}

\item{\code{basType}}{Object of class \code{"character"}. To be chosen from {"PCA", "B-splines"}.}

\item{\code{basis}}{Object of class \code{"list"}. Projection basis. For functional inputs, each element
(fDims_i x fpDims_i) contains the basis functions used for the projection of one functional input.}

\item{\code{coefs}}{Object of class \code{"list"}. Each element (n x fpDims_i) contains the coefficients used for
the projection of one functional input.}
}}

\author{
José Betancourt, François Bachoc, Thierry Klein and Jérémy Rohmer
}
