% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_SF.R
\name{setHypers_SF}
\alias{setHypers_SF}
\title{Optimization of hyperparameters for funGp models}
\usage{
setHypers_SF(sIn, fpIn, J, sMs, fMs, sOut, kerType, n.starts, n.presample)
}
\arguments{
\item{sIn}{a matrix of scalar input values to train the model. Each column must match an input variable and each row a training point.}

\item{fpIn}{a list with as many elements as functional inputs. The i-th element must be a matrix with the projection coefficients
for the i-th functional input.}

\item{J}{a list with as many elements as functional inputs. The i-th element must be the Gram matrix of the basis functions used for
the projection of the i-th functional input.}

\item{sMs}{a list with as many elements as scalar input variables. Each element of the list is a n times n matrix of differences
between the scalar observation coordinates.}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{a vector (or 1-column matrix) containing the values of the scalar output at the training points.}

\item{kerType}{a character string indicating the covariance structure to be used, to be choosen between "gauss", "matern5_2" or "matern3_2".}

\item{n.starts}{Fill!!!!!!!!!!}

\item{n.presample}{Fill!!!!!!!!!!}
}
\value{
The hyperparameters.
}
\description{
Sets good values for the hyperparameters of the Gaussian process model based on optimization.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
