% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_training_F.R
\name{setSPoints_F}
\alias{setSPoints_F}
\title{Setting up of starting points for optimization of hyperparameters}
\usage{
setSPoints_F(bnds, fMs, sOut, kerType, varfun, n.starts, n.presample, nugget)
}
\arguments{
\item{bnds}{Fill this!!!!!!!!!!}

\item{fMs}{a list with as many elements as functional input variables. Each element of the list is a n times n matrix of differences
between the functional observation coordinates.}

\item{sOut}{Fill this!!!!!!!!!!}

\item{kerType}{Fill this!!!!!!!!!!}

\item{niter}{Fill this!!!!!!!!!!}
}
\value{
Fill this!!!!!!!!!!
}
\description{
Fill this!!!!!!!!!!
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
\keyword{internal}
