% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGpModel_Class.R
\name{show}
\alias{show}
\alias{show,funGpModel-method}
\title{Fill!!!!!!!!!!!}
\usage{
\S4method{show}{funGpModel}(object)
}
\arguments{
\item{object}{An object to show.}
}
\description{
Fill!!!!!!!!!!!
}
