% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_show_Doc.R, R/2_fgpKern_Class.R,
%   R/2_fgpProj_Class.R, R/1_fgpm_Class.R, R/1_Xfgpm_Class.R
\name{show}
\alias{show}
\alias{show,fgpKern-method}
\alias{show,fgpProj-method}
\alias{show,fgpm-method}
\alias{show,Xfgpm-method}
\title{Printing methods for the funGp package}
\usage{
\S4method{show}{fgpKern}(object)

\S4method{show}{fgpProj}(object)

\S4method{show}{fgpm}(object)

\S4method{show}{Xfgpm}(object)
}
\arguments{
\item{object}{either a \linkS4class{fgpm}, \linkS4class{fgpKern}, \linkS4class{fgpProj}, or
\linkS4class{Xfgpm} object.}
}
\description{
This set of method enables printing of the main objects defined in the funGp package. That
  corresponds to \linkS4class{fgpm}, \linkS4class{fgpKern}, \linkS4class{fgpProj}, and \linkS4class{Xfgpm}
  objects, representing funGp models, data structures related to the kernel of the model, data structures
  related to projection of inputs, and structures related to structural optimization of the model,
  respectively.
}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
