% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funGp_Class.R
\name{simulate}
\alias{simulate}
\alias{simulate,funGp-method}
\title{Simulation method for the funGp Class}
\usage{
simulate(object, nsim = 1, seed = NULL, ...)

\S4method{simulate}{funGp}(
  object,
  nsim = 1,
  seed = NULL,
  sIn.sm = NULL,
  fIn.sm = NULL,
  nug.sim = 0,
  detail = "light",
  ...
)
}
\arguments{
\item{object}{An object to simulate from.}

\item{nsim}{fill!!!}

\item{seed}{fill!!!}

\item{...}{Further arguments for methods.}

\item{sIn.sm}{fill!!}

\item{fIn.sm}{fill!!}

\item{nug.sim}{fill!!!}

\item{detail}{fill!!!}
}
\description{
Simulation method for the funGp Class
}
\examples{
set.seed(100)
n.tr <- 25
sIn <- expand.grid(x1 = seq(0,1,length = sqrt(n.tr)), x2 = seq(0,1,length = sqrt(n.tr)))
fIn <- list(f1 = matrix(runif(n.tr*10), ncol = 10), matrix(runif(n.tr*22), ncol = 22))

# generating output data for training
sOut <- as.matrix(sapply(t(1:n.tr), function(i){
  x1 <- sIn[i,1]
  x2 <- sIn[i,2]
  f1 <- fIn[[1]][i,]
  f2 <- fIn[[2]][i,]
  t1 <- seq(0,1,length = length(f1))
  t2 <- seq(0,1,length = length(f2))
  as.numeric(x1 + 2 * x2 + 4 * mean(t1 * f1) + mean(f2))
}))

# creating a funGp model
m1 <- funGp(sIn = sIn, fIn = fIn, sOut = sOut)

# generating input data for simulation
set.seed(100)
n.sm <- 100
sIn.sm <- expand.grid(x1 = seq(0,1,length = sqrt(n.sm)), x2 = seq(0,1,length = sqrt(n.sm)))
fIn.sm <- list(f1 = matrix(runif(n.sm*10), ncol = 10), matrix(runif(n.sm*22), ncol = 22))

# making light simulations
m1.sims_l <- simulate(m1, nsim = 10, sIn.sm = sIn.sm, fIn.sm = fIn.sm)

# plotting light simulations
plotSims(m1, m1.sims_l)

# making full simulations
m1.sims_f <- simulate(m1, nsim = 10, sIn.sm = sIn.sm, fIn.sm = fIn.sm, detail = "full")

# plotting full simulations in full mode
plotSims(m1, m1.sims_f)

# plotting full simulations in light mode
plotSims(m1, m1.sims_f, detail = "light")

}
\author{
José Betancourt, François Bachoc and Thierry Klein
}
